requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
mcp
python-dotenv
httpx